ggn_checker/requirements.txt
# GGn eBook Checker Dependencies

# Core dependencies
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional for Excel support
openpyxl>=3.1.0
